requirements.txt
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
databricks-connect==13.3.0
databricks-sdk==0.18.0
exceptiongroup==1.2.0
google-auth==2.27.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
protobuf==4.25.2
py==1.11.0
py4j==0.10.9.7
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
tzdata==2023.4
urllib3==2.1.0
--extra-index-url https://databricks.github.io/workflows-authoring-toolkit
workflows_authoring_toolkit==0.2.0
setuptools
wheel
-e .
